drivers:platform:aducm3029: Remove #include <drivers/uart/adi_uart.h>
This change is related to
https://github.com/analogdevicesinc/EVAL-ADICUP3029/pull/49

In the PR is an adicup project that builds noos as a static library and a new project
that uses the library doesn't have this include in the include directory.

The irq_enable and disable implementation was changed to support this change.
