+++ drivers/platform/aducm3029/uart.c
@@ -41,6 +41,7 @@
 /************************* Include Files **************************************/
 /******************************************************************************/
 
+#include <drivers/uart/adi_uart.h>
 #include <stdlib.h>
 #include "uart.h"
 #include "irq.h"
@@ -68,6 +69,47 @@ struct baud_desc {
 	uint16_t	div_c;
 	/** From 0 to 3 */
 	uint8_t		osr;
+};
+
+/**
+ * @struct op_desc
+ * @brief It stores the state of a operation
+ */
+struct op_desc {
+	/** Is set when an write nonblocking operation is executing */
+	bool		is_nonblocking;
+	/** Current buffer*/
+	uint8_t		*buff;
+	/** Number of bytes pending to process */
+	uint32_t	pending;
+};
+
+/**
+ * @struct aducm_uart_desc
+ * @brief Stores specific parameter needed by the UART driver for internal
+ * operations
+ */
+struct aducm_uart_desc {
+	/** Handle needed by low level functions */
+	ADI_UART_HANDLE	uart_handler;
+	/** Stores the error occurred */
+	enum UART_ERROR	errors;
+	/** Set if callback is enabled */
+	bool		callback_enabled;
+	/**
+	 * Buffer needed by the ADI UART driver to operate.
+	 * This buffer allocated and aligned at runtime to 32 bits
+	 */
+	uint8_t		*adi_uart_buffer;
+	/**
+	 * Stores the offset used to align adi_uart_buffer.
+	 * Needed to deallocate \ref adi_uart_buffer
+	 */
+	uint32_t	adi_uart_buffer_offset;
+	/** Status of a write operation */
+	struct op_desc	write_desc;
+	/** Status of a read operation */
+	struct op_desc	read_desc;
 };
 
 /******************************************************************************/
@@ -186,7 +228,7 @@ static void uart_callback(void *ctx, uint32_t event, void *buff)
 			extra->read_desc.buff += len;
 		} else {
 			extra->read_desc.is_nonblocking = false;
-			if (extra->callback_enabled)
+			if (desc->callback)
 				desc->callback(desc->callback_ctx, READ_DONE,
 					       NULL);
 		}
@@ -204,7 +246,7 @@ static void uart_callback(void *ctx, uint32_t event, void *buff)
 			extra->write_desc.buff += len;
 		} else {
 			extra->write_desc.is_nonblocking = false;
-			if (extra->callback_enabled)
+			if (desc->callback)
 				desc->callback(desc->callback_ctx, WRITE_DONE,
 					       NULL);
 		}
@@ -213,7 +255,7 @@ static void uart_callback(void *ctx, uint32_t event, void *buff)
 		extra->errors |= (uint32_t)buff;
 		extra->read_desc.is_nonblocking = false;
 		extra->write_desc.is_nonblocking = false;
-		if (extra->callback_enabled)
+		if (desc->callback)
 			desc->callback(desc->callback_ctx, ERROR, buff);
 		break;
 	}
